Fix enumeration on windows
On Linux, enumeration looks like:
- Get Device Descriptor
  - SETUP
  - IN (only 1 packet of 8 bytes is read)
  - OUT (status, 0 bytes)
- Reset
- Set Address
- Get Device Descriptor

On Windows, it's the same without the Reset.

The problem is that when the usb-device crate wants to write the first IN packet of the second Get
Device Descriptor, it gets a WouldBlock because nrf-usbd still believes the second IN packet of the
first Get Device Descriptor is pending. But that's wrong, when we reached the OUT status of the
first Get Device Descriptor, any pending IN state should be discarded.

This PR fixes this by resetting the busy bit to its neutral state.

Tested on Windows and Linux.